Implemented quality comparison
The item objects' profit parameter is now calculated through the most profitable quality comparison. The parameter "highestProfitBMQuality" will have to be displayed in a column next to the BM columns so that the user knows which buy order in BM the sell orders are being compared to. The same goes for the sell order in the city columns (add a column for city quality).

The files "index_REMOVE_DUPLICATED_WORKING.html" and "index_testing.html" are temporary and will be removed before the merge with the master branch.